+++ node-manager/mindreader/mindreader.go
@@ -27,8 +27,8 @@ import (
 	"github.com/streamingfast/bstream/blockstream"
 	pbbstream "github.com/streamingfast/bstream/pb/sf/bstream/v1"
 	"github.com/streamingfast/dstore"
+	"github.com/streamingfast/firehose-core/internal/utils"
 	nodeManager "github.com/streamingfast/firehose-core/node-manager"
-	"github.com/streamingfast/firehose-core/utils"
 	"github.com/streamingfast/logging"
 	"github.com/streamingfast/shutter"
 	"go.uber.org/zap"
